refactor to create separate module for specific distributions
